Last push

+++ C05/ex05/ft_sqrt.c
@@ -6,7 +6,7 @@
 /*   By: mkoch-le <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/20 17:41:07 by mkoch-le          #+#    #+#             */
-/*   Updated: 2022/07/21 12:40:07 by mkoch-le         ###   ########.fr       */
+/*   Updated: 2022/07/23 18:58:36 by mkoch-le         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,11 +19,11 @@ int	ft_sqrt(int nb)
 	n = 1;
 	if (nb < 1)
 		return (0);
-	while (n * n < nb)
+	while (n * n < (unsigned int)nb)
 	{
 		n++;
 	}
-	if (n * n > nb)
+	if (n * n > (unsigned int)nb)
 		return (0);
 	return (n);
 }
